requirements.txt
streamlit
requests
beautifulsoup4
ibm-watsonx-ai
pypdf
python-dotenv
fastapi
uvicorn
pyngrok